+++ src/inspect_ai/tool/_tool_with.py
@@ -1,6 +1,4 @@
-import inspect
 from copy import copy
-from dataclasses import dataclass
 
 from inspect_ai._util.registry import (
     registry_info,
@@ -10,6 +8,8 @@ from inspect_ai._util.registry import (
 )
 
 from ._tool import Tool
+from ._tool_description import ToolDescription, set_tool_description
+from ._tool_info import parse_tool_info
 
 
 def tool_with(
@@ -29,15 +29,21 @@ def tool_with(
     Returns:
        A copy of the passed tool with the specified descriptive information.
     """
-    # validate that the parameters are all part of the tool's signature
+    # get the existing tool info
+    tool_info = parse_tool_info(tool)
+
+    # provide parameter overrides
     if parameters:
-        signature_param_names = inspect.signature(tool).parameters.keys()
+        signature_param_names = tool_info.parameters.properties.keys()
         for param_name in parameters.keys():
             if param_name not in signature_param_names:
                 raise ValueError(
                     f"tool_with error: no parameter named '{param_name}' "
                     + f"(valid parameters are {', '.join(signature_param_names)})"
                 )
+            tool_info.parameters.properties[param_name].description = parameters[
+                param_name
+            ]
 
     # copy the tool and set the descriptions on the new copy
     tool_copy = copy(tool)
@@ -45,24 +51,8 @@ def tool_with(
     set_registry_params(tool_copy, registry_params(tool))
     set_tool_description(
         tool_copy,
-        ToolDescription(name=name, description=description, parameters=parameters),
+        ToolDescription(
+            name=name, description=description, parameters=tool_info.parameters
+        ),
     )
     return tool_copy
-
-
-@dataclass
-class ToolDescription:
-    name: str | None = None
-    description: str | None = None
-    parameters: dict[str, str] | None = None
-
-
-def tool_description(tool: Tool) -> ToolDescription:
-    return getattr(tool, TOOL_DESCRIPTION, ToolDescription())
-
-
-def set_tool_description(tool: Tool, description: ToolDescription) -> None:
-    setattr(tool, TOOL_DESCRIPTION, description)
-
-
-TOOL_DESCRIPTION = "__TOOL_DESCRIPTION__"